Paco's requests
Added new sizing options for the applet.
The description now is shown before the applet (if something is written
there).
More html text can be written in a new text editor and the text will
appear after the applet (if something is written)
You will need to reinstall the whole plugin :-(